Bonus => Bonus<long>

+++ Assets/BrandChangeDescription.cs
@@ -12,5 +12,5 @@ public class BrandChangeDescription : UpgradedParameterView
         return "Brand power: " + (int)SelectedCompany.branding.BrandPower;
     }
 
-    Bonus<float> BrandPower => MarketingUtils.GetMonthlyBrandPowerChange(SelectedCompany, GameContext);
+    Bonus<long> BrandPower => MarketingUtils.GetMonthlyBrandPowerChange(SelectedCompany, GameContext);
 }
