+++ Assets/Utils/Companies/DesireToBuyOrSell/OpinionAboutAcquisition.cs
@@ -8,9 +8,10 @@ namespace Assets.Utils
     public static partial class CompanyUtils
     {
         // opinion about offer
-        public static Bonus<float> GetInvestorOpinionAboutAcquisitionOffer(AcquisitionOfferComponent ackOffer, GameEntity investor, GameEntity targetCompany, GameContext gameContext)
+        // TODO <float>
+        public static Bonus<long> GetInvestorOpinionAboutAcquisitionOffer(AcquisitionOfferComponent ackOffer, GameEntity investor, GameEntity targetCompany, GameContext gameContext)
         {
-            var container = new Bonus<float>("Opinion about acquisition offer");
+            var container = new Bonus<long>("Opinion about acquisition offer");
 
             switch (investor.shareholder.InvestorType)
             {
@@ -30,9 +31,9 @@ namespace Assets.Utils
 
         
 
-        public static Bonus<float> GetVentureOpinionAboutOffer(AcquisitionOfferComponent acquisitionOffer, GameEntity investor, GameEntity company, GameContext gameContext)
+        public static Bonus<long> GetVentureOpinionAboutOffer(AcquisitionOfferComponent acquisitionOffer, GameEntity investor, GameEntity company, GameContext gameContext)
         {
-            var bonus = new Bonus<float>("Venture investor Opinion");
+            var bonus = new Bonus<long>("Venture investor Opinion");
             var conditions = acquisitionOffer.BuyerOffer;
 
             var priceOk = conditions.Price < acquisitionOffer.SellerOffer.Price;
@@ -48,9 +49,9 @@ namespace Assets.Utils
             return bonus;
         }
 
-        public static Bonus<float> GetFounderOpinionAboutOffer(AcquisitionOfferComponent acquisitionOffer, GameEntity investor, GameEntity company, GameContext gameContext)
+        public static Bonus<long> GetFounderOpinionAboutOffer(AcquisitionOfferComponent acquisitionOffer, GameEntity investor, GameEntity company, GameContext gameContext)
         {
-            var bonus = new Bonus<float>("Founder Opinion");
+            var bonus = new Bonus<long>("Founder Opinion");
             var conditions = acquisitionOffer.BuyerOffer;
 
             var priceOk = conditions.Price < acquisitionOffer.SellerOffer.Price;
