+++ Assets/Utils/Marketing/BrandPower.cs
@@ -12,7 +12,8 @@ namespace Assets.Utils
             company.ReplaceBranding(brandPower);
         }
 
-        public static Bonus<float> GetMonthlyBrandPowerChange(GameEntity product, GameContext gameContext)
+        // float
+        public static Bonus<long> GetMonthlyBrandPowerChange(GameEntity product, GameContext gameContext)
         {
             var conceptStatus = ProductUtils.GetConceptStatus(product, gameContext);
 
@@ -29,7 +30,7 @@ namespace Assets.Utils
 
             var partnershipBonuses = GetPartnershipBonuses(product, gameContext);
 
-            var BrandingChangeBonus = new Bonus<float>("Brand power change")
+            var BrandingChangeBonus = new Bonus<long>("Brand power change")
                 //.AppendAndHideIfZero(percent + "% Decay", -(int)baseDecay)
                 .AppendAndHideIfZero("Is not paying for marketing", isNotPayingForMarketing ? -3 : 0)
                 .AppendAndHideIfZero("Outdated app", isOutOfMarket ? -3 : 0)
